chore: move grpc helper functions to GrpcHelper
