+++ src/meta/service/src/lib.rs
@@ -18,6 +18,7 @@
 pub mod api;
 pub mod configs;
 pub mod export;
+pub(crate) mod grpc_helper;
 pub mod message;
 pub mod meta_service;
 pub mod metrics;
